+++ src/__tests__/stringify-error-like.test.ts
@@ -1,0 +1,14 @@
+import { stringifyErrorLike } from '../stringify-error-like';
+
+describe(stringifyErrorLike.name, () => {
+	it('Stringifies an example', () => {
+		expect(
+			stringifyErrorLike({
+				message: 'Oops',
+				name: 'NaughtyError',
+				stack: 'at Foo\nat Bar',
+				code: 'TERRIBLE',
+			}),
+		).toBe(`NaughtyError: Oops (code=TERRIBLE). Stack:\nat Foo\nat Bar`);
+	});
+});