+++ src/error-like.ts
@@ -1,5 +1,5 @@
-export type ErrorLike = {
-	message: string;
-	stack: string;
+/** Type representing an object literal with all the same properties as an
+ * `Error` object plus an optional string "code" property */
+export interface ErrorLike extends Error {
 	code?: string;
-};
+}